refactor(interface-impl): move Node to structs.py to avoid import conflicts
